split AVS builder in 3 files
