+++ crates/avs/src/lib.rs
@@ -1,2 +1,6 @@
 //! Starts all the services for the AVS using futures
-pub mod builder;
+pub mod avs;
+/// Avs builder
+pub mod avs_builder;
+/// Launch Avs trait
+pub mod traits;
