+++ crates/avs/src/traits.rs
@@ -1,0 +1,7 @@
+use std::future::Future;
+
+/// Launch Avs trait
+pub trait LaunchAvs<T: Send + 'static> {
+    /// Launch Avs
+    fn launch_avs(self, avs: T) -> impl Future<Output = eyre::Result<()>> + Send;
+}
